improved MultiplyModPrime

+++ HashFunctions.cs
@@ -114,7 +114,8 @@ namespace HashFunctions {
             }
 
             // Efficient mod 2^l
-            uint bitmask = (uint) (1UL << (int) l) - 1;
+            ulong bitmask = (ulong) (1UL << (int) l) - 1;
+            
             return (ulong) (y & bitmask);
         }
